Re-enable newspeak cop
While I disagree with language changes that are intended purely for political
purpose and reduce clarity, it seems there is significant efforts being made to
make self-explanatory alternatives to historically used terms. This helps make
terms intuitive, for example allowlist/denylist is clear in purpose, while
whitelist/blacklist requires institutional knowledge.

If this deviates into territory that is less intuitive, I may return to
disabling this cop, but for now it's worth keeping on the radar at least.